# encoding/json: consolidate the isSpace function